[wlan] Use BSSID, not BSS desc. in SME deauth req.
Future CL fxrev.dev/914513 will require a minor change to the SME
function send_deauthenticate_request. Currently, that function takes the
BSS description as a parameter; but it only needs the BSSID. In certain
roaming failure scenarios, SME might not have the full BSS description,
but SME will have the BSSID. To reduce the cognitive load of that large
future CL, this CL simply changes the function to receive only what it
needs (the BSSID), with no change in logic.

Test: fx test wlan-sme-tests
Change-Id: I9a5a244935a2802f2746e49fb5d428239df90d05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080153
Commit-Queue: Karl Ward <karlward@google.com>
Reviewed-by: Dylan Swiggett <swiggett@google.com>