Reland "[debugger] Fix name resolution after dot operator"
The issue was that the breakpoint set in the script was set on a macro which gets inlined differently in different build configurations, so we wouldn't always end up in the right frame to print the `vec`. The solution is to set the breakpoint in a more predictable location that should always be present in all build supported build variants that pass the `optimize=none` check at the top of the script.

This is a reland of commit 97849a0aa29c56f6ad0c55e618a8ff2d6e6f7b75

Original change's description:
> [debugger] Fix name resolution after dot operator
>
> In specific situations where there is a type that conflicts with a
> member name, the expression parser will always take the type before
> looking for variable identifiers with that name.
>
> In the context of evaluating the right hand side of a "." or "->"
> operator in an expression, types should be ignored. If a type does
> happen to conflict with the token that would otherwise identify a
> particular member on an object, we should treat the name of the type
> like an identifier and then let the member accessor logic determine if
> there is actually an identifier there by that name.
>
> Fixed: 352125964
> Change-Id: I79538c6c82a8dbe23de58fbf95cc77563a419bf9
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079633
> Commit-Queue: Jacob Rutherford <jruthe@google.com>
> Reviewed-by: Adam Barth <abarth@google.com>

Change-Id: I3db306993a73673b936d6953974a78dd095f4e44
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081052
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>