[kernel][pi] Fix an old copy-pasta bug
Fix an old bug in an accessor used to access a thread's TID via a
thread adapter class in the scheduler PI code.

The `tid()` method was supposed to return the thread's TID, but
instead it was being used to return the thread's state instead.  The
method was being used only in one place, and it was for a
DEBUG_ASSERT_MSG which (by design) has not been firing, so we have
just never noticed.

It's fixed now.

Change-Id: I08d70acb9cbbf63013c46c2511955a3036c4db2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081976
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>