[memalloc] s/UpdateFreeRamSubranges/UpdateRamSubranges
When UpdateFreeRamSubranges() was introduced, we wished to effectively
reserve a special range (which could span both free RAM and existing
allocations), but wanted to do so in a non-destructive way: the result
was just to update the free RAM subranges leaving the existing
allocations in place. However, with allocations either 'surviving' into
the kernel proper or being freed at hand-off, this approach is no longer
suitable. We want the whole range with the intended allocated type to
be present in the kernel proper, and not with "free RAM" carveouts
resulting from the subranges that were originally occupied when we
called this method.

Accordingly, we destructively update the method to reallocate all RAM
ranges. These types are mostly for diagnostic purposes, and even though
such a destructive operation might cloud that observability, it should
remain clear enough what it going on in practice around any of these
callsites.

Bug: 347766366
Change-Id: I2af0a5c606015e729271d92e415dfadad07440dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079555
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>