[bazel-sdk] add build time check for restricted symbols
Adds the ability to provide a set of restricted symbols to be checked at
build time. If the library uses any of the restricted symbols then we
will fail at build time instead of runtime.

Note: We are not actually enabling this check in this CL until we
understand why symbols are being pulled in during RBE builds. Once the
linked bug is better understood we will enable this feature so we don't
regress in the future.

BUG: b/352586714
Change-Id: Icef189903c23cd531b307e8674a87ad057c89a45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081978
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>