[starnix][iptables] Separate namespace ID from table name
When creating a fuchsia.net.filter/Namespace to correspond to an
iptables table, starnix adds a prefix indicating the IP version ("ipv4-"
or "ipv6-") so that IPv4 tables don't conflict with IPv6 tables.
Refactor to clarify the distinction between the NamespaceId and the
table name. This fixes a bug in the logic that was using the prefixed
NamespaceId to determine routine type (IP or NAT), which meant that all
routines were determined to be IP routines.

Change-Id: Ie2f8b2db32be0c557f7022851f5aa276b6acaebc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1082712
Reviewed-by: David Song <wintermelons@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>