[bt][a2dp] Prevent lock over await
Perform release tasks on StreamEndpoint in the background after
responding to the peer.  Pass a future out of the handle_request
function which holds a lock on PeerInner, so we can await on the result,
instead of holding the lock while waiting.

This prevents the possibility of double-locking the peer if a reservation
was released while awaiting during a Close operation or while suspending
in response to a requested Start that could not be satisfied due to
a lack of resources.

Fixes: fxbug.dev/326585243
Test: updated unit tests
Test: fx test bt-a2dp-lib-tests bt-a2dp-tests bt-avdtp-tests
Change-Id: If676b2839f4d35c20ba866be578a04c9c7f1572d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1078733
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>