[libc] Improve debugdata test for failure cases
Preparing for a new implementation of __sanitizer_publish_data,
improve its tests.  The failure case was not quite tested
explicitly, only that it didn't make the helper crash.  Now the
helper knows whether it's a success test or a failure test and it
fails if the __sanitizer_publish_data appears to succeed when it
should have failed.  In addition to the test for no /svc in the
namespace, also add a test for an /svc whose channel peer is
closed so the channel send fails on a valid channel handle.

Bug: 342469121
Change-Id: I32d9f38ec6fdbb78735d818ff39b06f00aa771a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080354
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>