[msd-arm-mali] Keep lease while suspended
Currently, once the GPU is no longer active due to suspend, the power
lease is quickly discarded (because the current_gpu_powerdown_timeout_ is
set to 30 ms in the future). Once the system is resumed, the GPU will
remain suspended until some new GPU work triggers
PowerOnGpuForRunnableAtoms to re-acquire a lease.

To enable the GPU to restart work immediately when the system is
resumed, the lease must be held as long as there is runnable work in the
scheduler, even if the GPU isn't currently active.

Fixed: 351904457
Change-Id: I2b78265778c78c41b98e063c54787e7516c10b94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081932
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>