[kernel][vm] Multiple page request types
Refactors the page request system to support multiple different kinds
of outstanding page requests at time, with a new MultiPageRequest
object to manage them.

This also changes anonymous page requests to not attempt to pretend to
be a page source with a regular PageRequest, but be their own specific
AnonymousPageRequest.

As most code paths know exactly what kind of page request they might
produce, there is typically little to pass around a truly abstract
PageRequest.

Having a MultiPageRequest is not a benefit in the current code, but
this refactor is a required pre-requisite to supporting continuation
page requests.

Run-All-Tests: True
Change-Id: I0067246b144671d2a1b18c403647e15579614d5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1067753
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>