[flatland] Move FIDL bindings from HLCPP to Natural.
This CL focuses on the fuchsia.ui.composition.Flatland server-side
implementation, as well as the flatland unit tests.  Other APIs and
tests are out of scope.

Flatland shares some utility code with other sub-systems, such as the
UberStruct interface to the Flatland renderer.  To limit the scope of
this CL, we leave some uses of HLCPP, and convert as necessary with
`NaturalToHLCPP()` and `HLCPPToNatural()`.  Bugs have been filed to
ensure these are eventually cleaned up.

Multiply: flatland_unittests: 10
Fixed: 350823293
Change-Id: Ibeedb7840ce824c27a93f20d3d6a4cefb6c52f29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1072787
Commit-Queue: Josh Gargus <jjosh@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>