[kernel][vm] Remove redundant trim
Due to the way looking up multiple pages in a VmObject use to work,
that is the LookupPagesLocked interface, the requested pages from a
page fault was trimmed to prevent more than the actual faulting page
from being marked dirty. The current LookupCursor interface uses an
explicit IfExistPages lookup, which is not allowed to perform dirty
transitions, and so the requested length does not need specific
trimming.

This does not improve performance or change behavior, but it improves
code clarity as the existing comment, check and logic is now confusing
given the current LookupCursor behavior.

There are existing core-tests that explicitly validate that page faults
do not cause additional dirty transitions.

Change-Id: Ib344197eaba8c756cd63f2e947b897039ba0b5bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048655
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>