[product_bundle] Adding virtual devices to product bundle
This adds the virtual device definitions to the
product bundle by default if the board supports the emulator.

This enables using `ffx emu start` to start the image. However,
since some product bundles do not support networking, such as bringup,
ffx emu start should be used with the `--console` option since there
is no way to connect to the RemoteControlService on the device.

Bug: b/351921300
Change-Id: I82d5cc10d04eef41145ca8219f0537c44ae47c86
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081856
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>