Reland "[power][broker] Create internal lease element"
This is a reland of commit 53a4ea3a2583d15b686b2ef63173ea1efb81f636

The original commit broke the power e2e tests on the VIM3. When the
timer received the interrupt, it would grab a wake lease to keep
the system out of the suspended state and block the dispatcher until
the lease was satisfied. However, this change now does not consider
a lease satisfied until the consumer element acknowledges the level
update request, which it cannot do while the dispatcher is blocked.

In the near future, power APIs will be introduced to avoid the need
for consumer elements, thus eliminating this problem. The workaround
employed here is to not wait for the lease to reach satisfaction. In
the case where execution state is active, grabbing the lease will
activate the execution state claim and keep it active. In the case
where it is inactive, there is a possibility of this racing if the
hrtimer completes (thus ack'ing the interrupt) and the system tries
to suspend in the gap between the interrupt clearing and SAG running
its power elements once the interrupt handler returns. Given the
limited impact of this behavior and the line-of-sight to a better
fix, this cost is absorbed for now.

Original change's description:
> [power][broker] Create internal lease element
>
> Adds an intermediate, synthetic lease element that has an assertive
> dependency to the element being leased. This changes the behavior of
> leasing elements in the power broker to return 'Satisfied' only when the
> element being leased has reached the desired level. Prior to this
> change, the broker would return 'Satisfied' when the dependencies of the
> leased element have reached their required levels.
>
> Bug: 345829869
> Test: fx test src/power/broker
> Test: fx test fake-system-activity-governor-test
> Change-Id: I4ea1dbf58dda2bd778117bcaf69d1ebc5f62776d
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1066613
> Reviewed-by: Onath Dillinger <claridge@google.com>
> Reviewed-by: Michael Brunson <mbrunson@google.com>
> Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
> Reviewed-by: Caroline Liu <carolineliu@google.com>
> Commit-Queue: Prashanth Swaminathan <prashanthsw@google.com>

Bug: 345829869
Cq-Include-Trybots: luci.turquoise.global.try:begonia_eng.vim3-release-power
Change-Id: I08c270aa3ec83abec4f81c6c068b4c9dfa715000
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081854
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Reviewed-by: Caroline Liu <carolineliu@google.com>
Reviewed-by: Chris Masone <cmasone@google.com>
Commit-Queue: Prashanth Swaminathan <prashanthsw@google.com>
Fuchsia-Auto-Submit: Prashanth Swaminathan <prashanthsw@google.com>
Reviewed-by: Onath Dillinger <claridge@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>