[ffx] Reland dynamic adding of target to daemon
(This was reverted unnecessarily -- the culprit was another CL.)

OpenTarget(<addr>) should add the target if it doesn't already
exist in the TargetCollection. But the newly-added target wasn't
getting added with the specified address, so the OpenTarget actually
failed to find the target to ask it to establish the host-pipe.

Change-Id: I655a69817a992a858a7f54a878f0ca2e40287d25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1082638
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>