[versioning] Update API level bumping to use NEXT
In preparation for switching to "NEXT" as the in-development API level,
update the bumping and FIDL golden generation parts of the build to
always use "NEXT" as the in-development API level.

The old numbered in-development level model is still supported, and
golden files for `NEXT` are not yet generated. However, `bump_api_level`
now adds a `"supported"` API level and is thus no longer compatible with
the old model.

`fidl_api_compatibility_testing.md` now uses `NEXT` in the example and
will no longer be updated when bumping.

Bug: 326277078
Change-Id: I07a86ea640dcbd1d115a4118f490f8d737f841f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073566
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>