[netstack3] Allow Deferred Device Socket Removal
During deferred device removal, there is a point in time where
WeakDeviceIDs cannot be upgraded, but the device's state has not been
dropped. During this window, closing a device socket on the device would
lead to a panic, because the device state still holds a strong reference
to the device socket.

To fix this, add support to the DeviceSocketApi for deferred socket
removal.

Fixed: 353484978

Change-Id: I3d9e165c1c463868bfb3b62f20864d4ff39d7978
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1082054
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>