[starnix][selinux] Fix the composite policy framework initial SIDs
The set of initial SIDs is defined by the Reference Policy, and
the order of declarations is part of the ABI.

An "initial_sids" file containing the Reference Policy is added,
which must be passed as an explicit argument, separate from the
rest of the policy fragments. The policy merge process validates
that only the "initial_sids" file attempts to define new SIDs.

Bug: b/343891990
Change-Id: I54366de90a2bafd0ceb43b085587237fa78c326e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080952
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Wez <wez@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>