[dl] Tests add check with RTLD_NOLOAD
This CL adds checks in the beginning of tests to assert that the file
that will be dlopen-ed is not still loaded from a previous test.

`kSupportsNoLoadMode` is used to prevent `DlImpl` tests from dlopen-ing
with `RTLD_NOLOAD` until it supports it.

`kRetrievesFileWithNoLoad` is used so that `DlSystemTests`-Musl will
expect that files are retrieved even with `RTLD_NOLOAD` passed (see
https://fxbug.dev/348722959).

`kFile`/`kDepFile[N]` are used as test variables to store file names to
make it easier to replicate logic across tests.

Bug: 348722959, 348727901, 354043838
Multipy: dl-tests: 100, dl-unittests: 100
Change-Id: Ie699e0cfa8bbcb74df16d8bb45ae57eab35b4cc9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069674
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>