[assembly][bt] Correctly add bt-transport-uart driver to assembly
As is, the driver is only included on builds that require "Standard"
bluetooth (e.g. all Bluetooth components). It is valid for the driver to
be included on "Coreless" builds as well, so move up the platform bundle
inclusion before the early return in the config.

Test: Manual
Bug: 354005030

Change-Id: I954f720aa37fef8f7da7ba8de5cdf7b28c503830
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1084662
Reviewed-by: Jerry Belton <jcecil@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>