[driver_manager] Refactor for driver dynamic linking.
Pass in loader and bootstrap receiver to driver host runner.

We want the option to specify which loader is used for which
driver host.

The bootstrap receiver zx::channel will be used to communicate
with the driver host. Eventually this will become a proper protocol.

Bug: 341998660,49831408
Change-Id: I5ef2648ac89958391aa3a57f60f3a8949baf7dd7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1085000
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>