[kernel][sched] Have blocked threads temporarily contribute firm demand
This change introduces an optimization in which threads that become
blocked will contribute firm demand for the remainder of the activation
period in which that took place. This allows threads that quickly
unblock to complete the still-accounted-for firm work they set out to do
in that period and not be penalized, minimizing disruption.

Bug: 328641440
Change-Id: I6bf2131ab192dd94981668d4950e890b52a6e917
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029434
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>