[superproject] Roll llvm-project/libc [libc] Fix GPU 'printf' on strings with padding
Summary:
We get the `strlen` to know how much memory to allocate here, but it
wasn't taking into account if the padding was larger than the string
itself. This patch sets it to an empty string so we always add the
minimum size. This implementation is slightly wasteful with memory, but
I am not concerned with a few extra bytes here and there for some memory
that gets immediately free'd.
GitOrigin-RevId: 37df86339e80b2bb7fef732080bce178fd76f6d4
Original-Revision: d9f700708e866628d74a01d9716ff7eed7919f8f
Roller-URL: https://ci.chromium.org/b/8741821607133538369
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I9df24d6acf3931eb918e3e33bedb1853a5cf5917
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086155