[camera][ge2d] Add const to copy assignment
https://github.com/llvm/llvm-project/pull/89652 made
std::pair trivially copyable if its members are, and
it caused the following issue: "error: the parameter
for this explicitly-defaulted copy assignment operator
is const, but a member or base requires it to be
non-const". This patch adds const to copy assignment
operator to fix the issue.

Bug: 354627057
Change-Id: Ic3b4db830502c9a1ef5260e7b15dd24cb9ced783
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086536
Fuchsia-Auto-Submit: Gulfem Savrun Yeniceri <gulfem@google.com>
Reviewed-by: Matt Thiffault <mthiffault@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>