[tee] Wire up memory TEE calls
This implements the TEE basic memory management calls. The TEE memory
management APIs can be configured through hints to zero the contents of
new allocations and to keep allocations private from other TAs. We
always satisfy the second requirement and to minimize bookkeeping it is
simpler to always zero contents as well (OP-TEE does the same). To do
this we provide allocator configuration values. We have to provide two
different options depending on if the build is configured to use Scudo
(the default) or ASan's allocator.

Bug: b/332956721

Change-Id: I833f8be2be02b6d97080615e09778d4b75b092d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1068236
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aidan Wolter <awolter@google.com>