[kernel] Support boot timers in the kernel
Adds support for timers operating on the boot timeline to the kernel.
This required modifying the TimerQueue to contain two separate timer
lists: one for monotonic timers and one for boot timers.

The reference timeline a particular timer is operating on is declared on
construction of the timer object, and is encoded as an enum member
variable. This may be overhauled in future patches once stronger types
for time in the kernel have been developed.

This is a necessary prerequisite to pausing the monotonic clock, as the
resume timer is our first boot timer. However, because we do not
currently support generic boot timers (e.g., zx_timer_create does not
allow callers to specify the boot timeline), the code asserts that only
the resume timer is ever set.

Bug: 341785588, 328306129
Change-Id: Ia147db6eee00fcee339ee232790de405776b200d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080638
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>