[boot-shim] Introduce DevicetreeGenericWatchdogItem
This item will generate a generic watchdog item when any of the
registered watchdogs provides a matching string and is able to
successfully fill the payload.

A 'Watchdog' type must satisfy the API such that the generic item does
the matching and delegates the filling and internal characteristics of
the item.

Fixes: 352391159
Change-Id: I8ca7728335bec9c9ee0a800011831f54e764243c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1056873
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>