[tefmocheck] Add check for unable to resolve address.
This message can occur if we fail to connect to the target which can
happen before the test runs and thus would be unrelated to the test.
Adding a tefmo check will help us track this as well as avoid filing
bugs for tests that fail because of this.

Bug: 354707902
Change-Id: I54f2dcecdbc0d40493dce59887c0294d2f3ae935
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087272
Reviewed-by: Mahesh Saripalli <maheshsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>