[driver_manager] Beginnings of enabling dynamic linking of a driver.
- Currently only starting the driver host part has been implemented
in the node class, the loading of the driver has been left to the
next CL.

- use_dynamic_linking is added as an option in the driver's program.
Like use_next_vdso, opting into dynamic linking is a driver host
wide option, and the field will only be checked when colocate = false.

- To launch the driver host, the node will ask the
node manager / driver host runner to create a driver host,
which will call the driver host runner to create the component
and call the dynamic linker.

- The node manager / driver host runner takes in a dynamic linker
service factory. Currently the test just creates a new loader per
driver host, but in production this will be able to be configured
differently if desired.

Bug: 341998660
Multiply: driver-host-runner-tests :50

Change-Id: I5ca4bc0950189d6a8a748f3828ed8ab570313985
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1085012
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>