[serdedoc] Adding disconnected types to toc
Due to b/332348955 enums that have different variants are
not connected to the data type graph. As a result, the README.md
for the type is generated, but the _toc.yaml does not include the file.

This situation blocks running doc-checker on the generated reference
docs. As a stopgap, this change adds the disconnected types to the
bottom of the toc for assemblydoc.

Bug: b/332348955
Bug: b/344621741
Change-Id: I62587da4710ad6d09c86db2c182b63911e9bb8b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087694
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>