[merge_power] Normalize Data before Correlating
Suppose we have a signal and feature
```
signal: [5, 0, 0, 0, 0, 2, 2, 2, 2, 2]
feature: [10, 5, 5, 5, 5]
```

The feature has a high, followed by relative lows, so it intuitively
make sense to align it at the beginning.

But if we do the correlations
```
[5, 0, 0, 0, 0] o [10, 5, 5, 5, 5] = 50 + 0+ 0 + 0 + 0 = 50
```
And the other way:
```
[2,2,2,2,2] o [10, 5,5,5,5] = 20 + 10 + 10 + 10 + 10 = 60
```

It's better to align it with all the twos -- the medium values beat out the matching high values.

However, if we normalize first, we get
```
[1, 0, 0, 0, 0, .4, .4, .4, .4]
[1, 0, 0, 0 ,0]
````

Here, the feature now aligns correctly with the beginning of the
feature.

Change-Id: Ifdfecc6ba527ca7d5005936e517ffc9c398918b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086957
Reviewed-by: Chris Masone <cmasone@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>