[ffx] [pkg] store repo from CLI not daemon
The daemon should not make config changes. This changes one occurrence,
the setting of the repository.repositories config, so that it is updated
from the `ffx repository add/add-from-pm` plugins instead.

Also, because we are now writing the config from two places (the CLI and
the daemon), we need to make sure that the daemon picks up any changes
in the CLI by invalidating its cache. Otherwise, it may not see changes
made by the CLI, and it could  any updates. Otherwise the daemon can
overwrite the changes made by the CLI.

Bug: b/343293737
Change-Id: I7ea6acdf0ef8c9d64c43aefd85792d1d6ed69524
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060975
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>