[driver_runtime][rust] `Arena::into_raw` isn't unsafe
Remove the unsafe designation and warning from `Arena::into_raw`, since
it's not unsafe to leak the arena.

Also adds `fdf_arena_t` to the public api of the crate, since it's
exposed in some functions for going to and from raw pointers. It was
missed because it's just a type alias to `u32`, and so did not trigger a
warning.

Change-Id: I685bfaa06c8e6df87022800ddff608fd42765749
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087877
Reviewed-by: Charles Celerier <chcl@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>