[ffx] client-side discovery must set ssh port
When using client-side discovery, the port was not being set in
the daemon, meaning that it could only talk to a target with ssh
port 22. (In addition, the `ffx self-test` test
`test_get_ssh_address_includes_port()` assumed the port was always
22.)

Also, because this triggers an `ffx self-test` run, we need to
remove a test that had stopped working when client-side discovery
was introduced. (The test, `test_get_ssh_address_timeout()`, is
based on likely-obsolete functionality, but rather than also dealing
with the changed behavior with get-ssh-address timeouts in this CL,
I'm just removing it.)

Bug: 354273937
Change-Id: I276838db49c2c1b07867132a24cbc76db56f7e23
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086912
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>