[wlan][mlme] Add MLME FIDL -> Fullmac FIDL conversion functions
Implements functions that converts MLME FIDL to the corresponding
WlanFullmacImpl request type, with unit tests.

Unlike existing Banjo conversion code, the new conversion functions
do not truncate the length of vectors or arrays that are too long.
Instead, they return an Error that should be handled by the caller. This
avoids undefined and likely broken behavior if e.g. an SSID or RSNE list
is truncated during conversion.

This is not currently used in production, but will be used when
Banjo is removed from the Fullmac MLME.

Test: fx test wlan-fullmac-mlme
Change-Id: I3ac3b4d1606894ce4ff2218ea8c3736310a52b99
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1078737
API-Review: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: John Tan <johntan@google.com>