[vmo-blob] Remove readable event
In blobfs, a blob can be opened for reading before its done being
written. The event is used to signal when writing has finished and the
blob is now readable.

In fxblob, blobs can't be opened for reading before writing has finished
so the event isn't necessary.

Change-Id: Idd64f16f74f5e601b734afdd1dab198f4fe16821
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086598
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Nikita Jindal <nikitajindal@google.com>