[scenic] Use zx_cache_flush
Replaces usages of ZX_VMO_OP_CACHE_* operations with zx_cache_flush.
With RFC-0252 these VMO cache operations are incorrect as the VMO can
be a physical VMO. In most cases there is a mapping available and the
preferred zx_cache_flush can be used instead. For one case where a
mapping is not available a fallback to create one is added if a
physical VMO is detected.

Change-Id: I6fc4f5932d7e38136ee49352022cd53763486753
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1054132
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>