[wlan-frame-writer] Add support for writing to padded buffers
The write_frame_with_fixed_slice!() macro now supports filling the
beginning of the buffer with zeroes so a frame can be written with no
space at the end of the slice. This is being added in preparation for
network device buffers which will require this functionality.

Additionally write_frame_with_fixed_buffer was renamed to
write_frame_with_fixed_slice so that an arbitrary amount of space at
the back of a buffer could be excluded from a write by specifying
a slice.

Test: fx test wlan-frame-writer-tests
Change-Id: I3a362cd9aeaeb9b80cd014b58a71acf43c56279a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087178
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Kiet Tran <kiettran@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>